refactor: provide LumberjackService only through the provideLumberjack function
